+++ OOPHomework/Atribute.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace OOPHomework
+{
+    [AttributeUsage(AttributeTargets.Struct)]
+    class ExImAttribute : Attribute
+    {
+    }
+}
